extract CORS policy into new SRP core microcrate
