+++ Cargo.toml
@@ -57,6 +57,7 @@ members = [
   "crates/bitnet-eval-core",     # SRP: deterministic eval/scoring math helpers
   "crates/bitnet-feature-matrix",
   "crates/bitnet-client-ip-core", # SRP: reusable client IP extraction from forwarding headers
+  "crates/bitnet-cors-policy-core", # SRP: reusable CORS policy/header synthesis primitives
   "crates/bitnet-api-versioning-core", # SRP: reusable API version negotiation and routing helpers
   "crates/bitnet-api-key-auth-core", # SRP: API key auth policy/store primitives
   "crates/bitnet-server-health-types-core", # SRP: reusable server health endpoint contract types
@@ -247,6 +248,7 @@ bitnet-common = { path = "crates/bitnet-common", version = "0.2.1-dev" }
 bitnet-math = { path = "crates/bitnet-math", version = "0.2.1-dev" }
 bitnet-warn-once = { path = "crates/bitnet-warn-once", version = "0.2.1-dev" }
 bitnet-atomic-file-core = { path = "crates/bitnet-atomic-file-core", version = "0.2.1-dev" }
+bitnet-cors-policy-core = { path = "crates/bitnet-cors-policy-core", version = "0.2.1-dev" }
 bitnet-tokenizer-text-core = { path = "crates/bitnet-tokenizer-text-core", version = "0.2.1-dev" }
 bitnet-models = { path = "crates/bitnet-models", version = "0.2.1-dev" }
 bitnet-quantization = { path = "crates/bitnet-quantization", version = "0.2.1-dev", default-features = false }
@@ -434,6 +436,7 @@ bitnet-quantization-bits = { path = "crates/bitnet-quantization-bits", version =
 bitnet-warn-once = { path = "crates/bitnet-warn-once", version = "0.2.1-dev" }
 bitnet-thread-config-core = { path = "crates/bitnet-thread-config-core", version = "0.2.1-dev" }
 bitnet-atomic-file-core = { path = "crates/bitnet-atomic-file-core", version = "0.2.1-dev" }
+bitnet-cors-policy-core = { path = "crates/bitnet-cors-policy-core", version = "0.2.1-dev" }
 # Core dependencies
 anyhow = "1.0.100"
 log = "0.4.28"
