+++ crates/bitnet-cors-policy-core/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "bitnet-cors-policy-core"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP microcrate for CORS policy configuration and header synthesis"
+
+[lib]
+path = "src/lib.rs"
+
+[lints]
+workspace = true
