+++ crates/bitnet-server/Cargo.toml
@@ -35,6 +35,7 @@ bitnet-request-context-core = { path = "../bitnet-request-context-core", version
 bitnet-api-versioning-core = { path = "../bitnet-api-versioning-core", version = "0.2.1-dev" }
 bitnet-server-health-types-core = { path = "../bitnet-server-health-types-core", version = "0.2.1-dev" }
 bitnet-client-ip-core = { path = "../bitnet-client-ip-core", version = "0.2.1-dev" }
+bitnet-cors-policy-core = { path = "../bitnet-cors-policy-core", version = "0.2.1-dev" }
 bitnet-startup-contract-guard = { path = "../bitnet-startup-contract-guard", version = "0.2.1-dev" }
 chrono.workspace = true
 clap.workspace = true
